Add links whenever we create a new trace and we have one ongoing
We often go from logs to traces, and to do that we use the trace ID.
However, sometimes the trace ID leads to empty traces like the following one:
https://tracing.cfdata.org/trace/96133b1ba9b23a2bb5000132501be3dd
This happens because the trace ID logging tag is only set on one place:
the connection ingress. However, the trace ID is often overridden later
(particularly when we're stitching the service's traces with some trace
ID that's coming from a downstream system).

We could change the logged ID, but that means we'd lose context.
Instead, stitching the old trace with the new trace (much like we do in
the connection vs request traces) seems like a superior option since it
allows us to easily navigate between the multiple related traces. That's
what this PR does.