Removes cyclical dependency
Causes issues with publishing. `bedrock` depends on `bedrock-macros` so
`bedrock-macros` is published first. However `bedrock-macros` has a
dev-dependency on `bedrock`.

Luckily for us, its not actually needed so we can just remove it.

However I think it's actually a valid setup, so in the context of the
private registry, we may run into this again.