telemetry: add tokio runtime metrics (#12)
This PR introduces a new unstable API for collecting tokio runtime
metrics which is gated behind `--cfg tokio_unstable` and `--cfg
foundations_unstable`.

This doesn't add all metrics. In particular, tokio's poll duration
histogram is not present since some additions might need to be made to
our histogram API.

More documentation will come in the future for some of this as we get a feel for whether we like these APIs or not.