ZTC-1478: Rework logger nesting tracking
This is a rework of commit 341befe7cd201fad4e39d60e5c12b03861a934d7. I did
that commit in the way that I did it, because I was trying to keep the
externally visible API the same. There was a public function, slog_logger(),
which promised to return an Arc<RwLock<Logger>>. So that meant we had to have
an Arc<RwLock<Logger>> lying around. This constrained the design and made
things grosser than they might otherwise have been.

It has been pointed out that we're already in the situation where the next
release we do is going to need to be a major version bump, anyway. So we might
as well revisit this, and do it the less-gross way that (slightly) breaks API
compatibility.

Other than that, no functionality has been changed from what was put in place
by commit 341befe7cd201fad4e39d60e5c12b03861a934d7.