ZTC-1545: Error during settings deserialization if YAML contains unused keys (#49)
The goal is to help users of foundations catch possible configuration
mistakes which, depending on the context, could have serious
consequences. For example, if a config field is renamed, but the name is
not changed in the deserialized YAML file, foundations will use the
defaults for the renamed config value, which may differ substantially
from the previous value. With this change, if the feature is enabled,
foundations will return an error on deserializing the YAML file since it
contains an unused key.

This takes a different approach from #43. This PR results in an error
instead of a logged warning, making it much harder to ignore.

If the opt-in foundations feature `settings-deny-unknown-fields` is
enabled, all settings containers using the `#[settings]` macro will have
the `#[serde(deny_unknown_fields)]` attribute set on them.

---------

Co-authored-by: Jackson Kruger <jkruger@cloudflare.com>