Expose active traces through telemetry server
Tracks currently-active spans in a low-overhead manner (relative to the
general overhead of tracing, at least). This is particularly useful because
even unsampled traces are available to be viewed here.

The model for this functionality is https://pkg.go.dev/golang.org/x/net/trace
although there is no built-in viewer here, we expect you to use Chrome's
about:tracing or anything that supports the same JSON log format.