More expressive error for EMSGSIZE from Jaeger reporter
We observe around 22k of these errors per day in production and don't
have any information to investigate further. By adding the span name as
well as some tag and log information, we should be able to pin down
which trace spans are excessively large and limit those in our code.