Re-export Hyper types used in public API
Foundations uses Hyper types in it's public API for
`TelemetryServerRoute` which means any users will need to add an
explicit dependency on the same version of hyper and ensure it stays
matching. Re-exporting the types will allow users to explicitly use the
same types as foundations is using.

For simplicity, we re-export the entire `hyper` crate under a new
`foundations::telemetry::reexports` module.