Remove default constraint from `Settings` trait
This required that fields of structs that derive `Settings` implement
`Default`, even if a default implementation of the field was provided.

This occurred in particular when trying to use a `NonZero` type within a
settings struct:

```
pub struct MySettings {
    #[serde(default = "MySettings::default_my_field")]
    pub my_field: NonZeroUsize,
}

impl MySettings {
    fn default_my_field() -> NonZeroUsize {
        return NonZeroUsize::new(1).unwrap()
    }
}
```

Instead, this change moves the `Default` constraint to only where the
`Settings` is used with a `Cli`.