Fix rerun directive issuance for build inputs
561158b changed build.rs to use CargoCallbacks with the option to issue
the rerun directive for all input headers. This caused unconditional
rebuilds of foundations because one of the input headers is generated in
build.rs.

While 9479fad prevents unconditional rebuilds, it also prevents the
rerun directive from being issued for any input header, which can be
annoying during development.

This change fine tunes rerun directive issuance, allowing it to be
issued for input headers but excluding all generated files (which we
assume are all files under `out_dir`).
We also now issue the rerun directive for the `seccomp.h` header template.