Don't use `[!Default; 0]: Default` impl (#150)
There is an effort to remove Rust's unconditional `[T; 0]: Default` impl
for `T: !Default`, see https://github.com/rust-lang/rust/pull/145457.
foundations depends on this impl in its public API, so fixing this is a
breaking change.

However, after consideration we decided that we will break semver for
this change. We don't expect any users of foundations to depend directly
on `[T; 0]: Settings`. We also put a best-effort workaround in place for the
`#[derive(Settings)]` use case inside foundations.

Co-authored-by: Leo Blöcher <lblocher@cloudflare.com>