Add `unprefixed` flag to `#[metrics]` macro
This change is twofold. First, it refactors the registry selection in
`metrics::internal::Registries` to use boolean flags rather than
separate methods to select a registry to use. As part of this, we add a
combination of flags to return a registry without the global service
name prefix that is normally present.

Second, we add a new `unprefixed` argument to the `metrics` proc-macro.
If present, this argument causes all metrics in the associated module to
be registered in the registry mentioned above. This works for both the
"main" and the "optional" registry.

I've added both a proc-macro expansion test and an integration test
to validate the flag's functionality.