Remove globally-shared INACTIVE_SPAN RwLock
This RwLock was a heavy contention point in applications that generate a
lot of spans, as every span operation must lock and unlock the RwLock.
Each lock/unlock is an atomic RMW and thus dirties the cache line for
all other cores.

We can avoid the global singleton `INACTIVE_SPAN` entirely with some
changes to our internal `SharedSpanHandle` API.