Fixup ratelimiter benchmarking
Previously the `iters` was divided amongst threads, meaning when the
number of threads increased, the amount of work per-thread decreased.
This had the fun effect of causing our high contention benchmark to
report ~3ns per iter...

Now every thread does `iters` amount of work and we see expected (?)
performance under contention. When we're below the number of CPUs on
my device, I see ~30ns per-iter which is what we get for the uncontended
benchmark. When we use 16 threads (more than the number of CPUs), the
threads start competing for CPU time and we jump to ~45ns per-iter.